Stats:  Reset after every display and make them fixed width
The help for --stats-every says
    "Show and reset stats every <seconds> seconds"
but the periodic stats were only resetting every 60 seconds with the
current stats.  So as to not disturb current_stats and the json output,
when the display interval is less than the reset interval, the
_difference_ between current and periodic will now be displayed.  The
behavior when the display interval is equal to or greater than the
reset interval is unchanged.

NOTE:  In either case, if the display interval is not an even divisor
into, or an even multiple of, the reset interval, the stats displayed
can occasionally be less because the current stats were reset in the
middle of a display interval.  This has always been the case however.

The numbers displayed ion the stats output were also very hard to
keep track of because the display widths kept changing depending
on the value.  They are now displayed with fixed widths.